# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AutoList"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/playlist-items/{item_id}/backdrop": {
      "get": {
        "description": "影片详情横幅：fanart.tv 剧照优先，没有时用 TMDB 剧照（w1280）。",
        "operationId": "playlist_item_backdrop_api_playlist_items__item_id__backdrop_get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "v",
            "required": false,
            "schema": {
              "default": "",
              "title": "V",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Playlist Item Backdrop"
      }
    },
    "/api/playlist-items/{item_id}/fanart-poster": {
      "get": {
        "description": "Proxy a fanart.tv poster; when fanart.tv has none, serve the Emby/TMDB poster instead.\n\n只有地址里的版本号与当前海报一致时才允许浏览器长期缓存；“待定”地址返回的是本次刚选出的海报，\n之后列表会换成带版本号的地址，因此只短期缓存，重新挑选海报后不会继续显示旧图。",
        "operationId": "playlist_item_fanart_poster_api_playlist_items__item_id__fanart_poster_get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "v",
            "required": false,
            "schema": {
              "default": "",
              "title": "V",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Playlist Item Fanart Poster"
      }
    },
    "/api/playlist-items/{item_id}/tmdb-poster": {
      "get": {
        "description": "Proxy a TMDB poster so films outside Emby still get artwork without exposing the TMDB key.",
        "operationId": "playlist_item_tmdb_poster_api_playlist_items__item_id__tmdb_poster_get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Playlist Item Tmdb Poster"
      }
    },
...
  }
}